Update userInterface.ino
changed data type to float

+++ masterSAM/userInterface.ino
@@ -58,7 +58,7 @@ void lcdPrint(int state, int dataState = noDatFlag)
   //lcd.print(line[3]); //perhaps could use fourth line to display target psi?
 }
 
-void sdWrite(int val) 
+void sdWrite(float val) 
 {
   if (myFile) {
     myFile.println(val);
